extract `parseQuery` function to its own file
